requirements.txt
aiohttp-socks
rich-argparse
aiohttp~=3.8.4
aiofiles~=23.1.0
tqdm~=4.65.0
aiolimiter~=1.1.0
python-dotenv~=1.0.0
